#15 divided auth router
