+++ server/routes/auth/login.js
@@ -1,5 +1,4 @@
 import jwt from "jsonwebtoken";
-import { promisify } from "util";
 import { redisClient } from "../../utils/redis.js";
 
 import User from "../../models/user.js";
@@ -94,38 +93,4 @@ const login = (req, res) => {
     .catch(onError);
 };
 
-/*
-    GET /api/auth/verify
-*/
-
-const verify = (req, res) => {
-  res.json({
-    success: true,
-    info: req.decoded,
-  });
-};
-
-// TO DO : refresh router
-const refresh = () => {
-  const secret = req.app.get("jwt-secret");
-  const options = JSON.parse(process.env.REFRESH_OPTIONS);
-
-  return jwt.sign({}, secret, options);
-};
-
-const refreshVerify = async (token, username) => {
-  const getAsync = promisify(redisClient.get).bind(redisClient);
-  const SUCCESS_RESULT = { ok: true };
-  const FAIL_RESULT = { ok: false };
-  try {
-    const data = await getAsync(username);
-    if (token === data) {
-      return SUCCESS_RESULT;
-    } else {
-      return FAIL_RESULT;
-    }
-  } catch (error) {
-    return FAIL_RESULT;
-  }
-};
-export { login, verify, refresh, refreshVerify };
+export default login;
