+++ server/routes/auth/refresh.js
@@ -1,0 +1,9 @@
+import jwt from "jsonwebtoken";
+
+const refresh = () => {
+  const secret = req.app.get("jwt-secret");
+  const options = JSON.parse(process.env.REFRESH_OPTIONS);
+
+  return jwt.sign({}, secret, options);
+};
+export default refresh;
