+++ server/routes/auth/verify.js
@@ -1,0 +1,12 @@
+/*
+    GET /api/auth/verify
+*/
+
+const verify = (req, res) => {
+  res.json({
+    success: true,
+    info: req.decoded,
+  });
+};
+
+export default verify;
